+++ clinicadl/predictor/predictor.py
@@ -8,48 +8,13 @@ from clinicadl.data.datasets import CapsDataset
 from clinicadl.data.utils import tsv_to_df
 from clinicadl.dictionary.words import GROUPS, PARTICIPANT_ID
 from clinicadl.experiment_manager import ExperimentManager
-from clinicadl.experiment_manager.data_group import DataGroup
+from clinicadl.experiment_manager.maps_reader import MapsReader
 from clinicadl.model import ClinicaDLModel
 from clinicadl.splitter.split import Split
 from clinicadl.utils.exceptions import (
     ClinicaDLConfigurationError,
     ClinicaDLDataLeakageError,
 )
-
-
-class MapsReader:
-    maps_path: Path
-
-    def _create_data_group(
-        self, name: str, caps_dataset: CapsDataset, split: Optional[int] = None
-    ) -> DataGroup:
-        """
-        Check that a data_group is not already written and writes the characteristics of the data group
-        (TSV file with a list of participant / session + JSON file containing the CAPS and the preprocessing).
-        """
-        data_group = DataGroup(name=name, split=split, maps_path=self.maps_path)
-        if data_group.exists():
-            raise ClinicaDLConfigurationError(
-                f"Data group {data_group.name} already exists, please give another name to your data group"
-            )
-
-        data_group.create(caps_dataset)
-        return data_group
-
-    def _load_data_group(self, name: str, split: Optional[int] = None) -> DataGroup:
-        """creates a new data_group."""
-        data_group = DataGroup(name=name, split=split, maps_path=self.maps_path)
-        if data_group.exists():
-            return data_group
-
-        raise ClinicaDLConfigurationError(
-            f"Could not find data group {data_group.name}"
-        )
-
-    def get_train_val_df(self):
-        """Loads the train and validation data groups."""
-        path = self.maps_path / GROUPS / "train+validation.tsv"
-        return tsv_to_df(path)
 
 
 class Predictor:
